adaptive methodist pages
- добавила библиотеку classNames
- разделила логику для мобильный и десктопной версии страниц:
-- TrainerAnalyticsPage, MainPage, EvaluationTablePage, DetailedAnalyticsPage
- добавила компонент Table
- изменила частично брейкпоинты
- добавила мок-данные для тренеров
