+++ frontend/package.json
@@ -6,6 +6,7 @@
 		"@nivo/line": "~0.87.0",
 		"@reduxjs/toolkit": "~2.2.3",
 		"axios": "~1.7.2",
+		"classnames": "^2.5.1",
 		"js-cookie": "~3.0.5",
 		"moment": "~2.30.1",
 		"react": "~18.3.0",
