+++ frontend/src/app/styles/_mixins.scss
@@ -23,13 +23,25 @@
 }
 
 @mixin lg() {
-	@media (max-width: 1440px) {
+	@media (max-width: 834px) {  // tablet vertical
 		@content;
 	}
 }
 
 @mixin xl() {
-	@media (max-width: 1920px) {
+	@media (max-width: 1024px) {
+		@content;
+	}
+}
+
+@mixin xxl() {
+	@media (max-width: 1200px) {
+		@content;
+	}
+}
+
+@mixin xxxl() {
+	@media (max-width: 1400px) {
 		@content;
 	}
 }
