--- frontend/src/pages/supervisor/Analytics.view.page.module.scss
@@ -1,29 +1,0 @@
-@use '~styles/mixins' as *;
-
-
-.list {
-	display: flex;
-	flex-direction: column;
-	padding: 25px 50px;
-	gap: 20px;
-	li {
-		list-style: disc;
-	}
-}
-
-.comment {
-	justify-content: center;
-	cursor: pointer;
-
-	@include sm() {
-		max-width: 163px ;
-		height: 210px;
-		text-align: center;
-	}
-}
-
-.buttons {
-	display: flex;
-	flex-direction: column;
-	gap: 20px;
-}
