+++ frontend/src/pages/supervisor/DetailedAnalyticsPage/ui/DetailedAnalyticsPage.jsx
@@ -1,6 +1,6 @@
 import useResize from '@shared/hooks/useResize';
-import { DetailedAnalyticsPageContent as DesktopContent } from './desktop/DetailedAnalyticsPageContent';
-import { DetailedAnalyticsPageContent as MobileContent } from './mobile/DetailedAnalyticsPageContent';
+import { DesktopContent } from './desktop/DesktopContent';
+import { MobileContent } from './mobile/MobileContent';
 
 export const DetailedAnalyticsPage = () => {
 	const isMobile = useResize('sm');
