+++ frontend/src/pages/supervisor/DetailedAnalyticsPage/ui/desktop/DesktopContent.module.scss
