+++ frontend/src/pages/supervisor/EvaluationTablePage/ui/EvaluationTablePage.jsx
@@ -1,0 +1,8 @@
+import useResize from '@shared/hooks/useResize';
+import { DesktopTableContent } from './decktop/DesktopTableContent';
+import { MobileTableContent } from './mobile/MobileTableContent';
+
+export const EvaluationTablePage = () => {
+	const isMobile = useResize('sm');
+	return <>{isMobile ? <MobileTableContent /> : <DesktopTableContent />}</>;
+};
