+++ frontend/src/pages/supervisor/EvaluationTablePage/ui/mobile/MobileTableContent.module.scss
@@ -1,0 +1,10 @@
+@use '~styles/mixins' as *;
+
+@include sm() {
+	.btn_wrap {
+		width: 100%;
+		display: flex;
+		flex-direction: column;
+		gap: 20px;
+	}
+}
