+++ frontend/src/pages/supervisor/TrainerAnalyticsPage/ui/decktop/AnalyticsContent.module.scss
