+++ frontend/src/shared/const/routes/supervisor.js
@@ -1,12 +1,12 @@
 import { MainPage } from '@supervisor/MainPage/ui/MainPage';
 import { TrainerAnalyticsPage } from '@supervisor/TrainerAnalyticsPage/ui/TrainerAnalyticsPage';
 import { DetailedAnalyticsPage } from '@supervisor/DetailedAnalyticsPage/ui/DetailedAnalyticsPage';
+import { EvaluationTablePage } from '@supervisor/EvaluationTablePage/ui/EvaluationTablePage';
 import ProgressGraphPage from '@supervisor/Progress.graph.page';
 import ProgressGraphViewPage from '@supervisor/Progress.graph.view.page';
 import SchedulePage from '@supervisor/Schedule.page';
 import CreateSchedulePage from '@supervisor/Create.schedule.page';
 import FeedbackPage from '@supervisor/Feedback.page';
-import EvaluationTable from '@supervisor/Evaluation.table.page';
 import NotificationPage from '@supervisor/Notification.page';
 import analytics from '../../assets/svg/sidebar/analytics.svg';
 import progress from '../../assets/svg/sidebar/progress.svg';
@@ -43,8 +43,8 @@ const supervisorRouter = [
 		element: <DetailedAnalyticsPage />
 	},
 	{
-		path: '/evaluation',
-		element: <EvaluationTable />,
+		path: '/evaluation', //заменить
+		element: <EvaluationTablePage />,
 		local: 'Таблицы',
 		img: tables
 	},
