+++ frontend/src/widgets/table/index.js
@@ -1,0 +1,1 @@
+export { Table } from './ui/Table';
