+++ frontend/src/widgets/table/ui/Table.module.scss
@@ -9,6 +9,15 @@
 	font-weight: var(--fw-regular);
 	text-align: center;
 	border: 0.5px solid #8f8f8f;
+
+	@include lg() {
+		overflow-x: auto;
+		font-size: 16px;
+	}
+	@include sm() {
+		font-size: 12px;
+		grid-template-columns: 50px 126px 1fr 1fr 65px 114px;
+	}
 }
 
 .items {
@@ -36,11 +45,11 @@
 	font-size: var(--fz-btn);
 	font-family: var(--ff-first);
 	font-weight: var(--fw-medium);
-}
 
-.btn_wrap {
-	display: flex;
-	flex-direction: column;
-	gap: 20px;
-	margin-left: auto;
+	@include md() {
+		font-size: 18px;
+	}
+	@include sm() {
+		font-size: 12px;
+	}
 }
